Safari-609.1.11.1/LayoutTests/fast/events/pageshow-pagehide-on-back-cached-expected.txt

ALERT: window.onload
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: line 25: Use of window.alert is not allowed while unloading a page.
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = true
Test pageshow/pagehide event behavior when navigating back to an cached page.
